requirements/requirements-base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-base.in
#
bcrypt==3.2.0
    # via -r requirements-base.in
cffi==1.14.5
    # via bcrypt
click==7.1.2
    # via uvicorn
ecdsa==0.14.1
    # via python-jose
fastapi==0.63.0
    # via -r requirements-base.in
h11==0.12.0
    # via uvicorn
passlib==1.7.4
    # via -r requirements-base.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.20
    # via cffi
pydantic==1.7.3
    # via fastapi
pymysql==0.9.3
    # via -r requirements-base.in
python-jose==3.2.0
    # via -r requirements-base.in
rsa==4.7.1
    # via python-jose
six==1.15.0
    # via
    #   bcrypt
    #   ecdsa
    #   python-jose
sqlalchemy==1.3.23
    # via -r requirements-base.in
starlette==0.13.6
    # via fastapi
uvicorn==0.13.3
    # via -r requirements-base.in
